requirements.txt
django==4.1.1
Pillow==9.3.0
django-ckeditor==6.5.1